Updated stop_container.sh for resolving same issue that is stopping container,

+++ scripts/stop_container.sh
@@ -1,10 +1,21 @@
-#!/bin/bash
-set -e
+# #!/bin/bash
+# set -e
 
-# Stop the running container (if any)
-# echo "Hi"
+# # Stop the running container (if any)
+# # echo "Hi"
+
+# echo "Container deletion initiated!!"
+# containerid=`docker ps | awk -f "" '{print $1}'`
+# docker rm -f $containerid
+# echo "Container deleted successfully!!"
+
+#!/bin/bash
 
 echo "Container deletion initiated!!"
-containerid=`docker ps | awk -f "" '{print $1}'`
-docker rm -f $containerid
-echo "Container deleted successfully!!"
+containerid=$(docker ps -q)
+if [ -n "$containerid" ]; then
+    docker rm -f $containerid
+    echo "Container deleted successfully!!"
+else
+    echo "No running containers found to stop."
+fi
